Commitment key cpped
