+++ cpp/src/barretenberg/proof_system/pcs/commitment_key.hpp
@@ -43,62 +43,72 @@ struct Params {
     class CommitmentKey {
 
       public:
+        barretenberg::scalar_multiplication::pippenger_runtime_state<curve::BN254> pippenger_runtime_state;
+        std::shared_ptr<barretenberg::srs::factories::ProverCrs<curve::BN254>> srs;
+
         CommitmentKey() = delete;
 
-        /**
-         * @brief Construct a new Kate Commitment Key object from existing SRS
-         *
-         * @param n
-         * @param path
-         *
-         */
         CommitmentKey(const size_t num_points, std::shared_ptr<barretenberg::srs::factories::CrsFactory> crs_factory);
 
-        // Note: This constructor is used only by Plonk; For Honk the CommitmentKey is solely responsible for extracting
-        // the srs.
         CommitmentKey(const size_t num_points,
                       std::shared_ptr<barretenberg::srs::factories::ProverCrs<curve::BN254>> prover_srs);
 
-        /**
-         * @brief Uses the ProverSRS to create a commitment to p(X)
-         *
-         * @param polynomial a univariate polynomial p(X) = ∑ᵢ aᵢ⋅Xⁱ ()
-         * @return Commitment computed as C = [p(x)] = ∑ᵢ aᵢ⋅[xⁱ]₁ where x is the secret trapdoor
-         */
         Commitment commit(std::span<const Fr> polynomial);
-
-        barretenberg::scalar_multiplication::pippenger_runtime_state<curve::BN254> pippenger_runtime_state;
-        std::shared_ptr<barretenberg::srs::factories::ProverCrs<curve::BN254>> srs;
     };
 
     class VerificationKey {
 
       public:
+        std::shared_ptr<barretenberg::srs::factories::VerifierCrs> verifier_srs;
+
         VerificationKey() = delete;
 
-        /**
-         * @brief Construct a new Kate Verification Key object from existing SRS
-         *
-         * @param num_points
-         * @param verifier_srs verifier G2 point
-         */
         VerificationKey([[maybe_unused]] size_t num_points,
                         std::shared_ptr<barretenberg::srs::factories::CrsFactory> crs_factory);
 
-        /**
-         * @brief verifies a pairing equation over 2 points using the verifier SRS
-         *
-         * @param p0 = P₀
-         * @param p1 = P₁
-         * @return e(P₀,[1]₁)e(P₁,[x]₂) ≡ [1]ₜ
-         */
         bool pairing_check(const GroupElement& p0, const GroupElement& p1);
-
-        std::shared_ptr<barretenberg::srs::factories::VerifierCrs> verifier_srs;
     };
 };
 
 } // namespace kzg
+
+namespace ipa {
+
+struct Params {
+    using Fr = typename barretenberg::g1::Fr;
+    using Commitment = typename barretenberg::g1::affine_element;
+    using GroupElement = barretenberg::g1::element;
+
+    using Polynomial = barretenberg::Polynomial<Fr>;
+
+    class CommitmentKey;
+    class VerificationKey;
+
+    class CommitmentKey {
+
+      public:
+        barretenberg::scalar_multiplication::pippenger_runtime_state<curve::BN254> pippenger_runtime_state;
+        std::shared_ptr<barretenberg::srs::factories::ProverCrs<curve::BN254>> srs;
+
+        CommitmentKey() = delete;
+
+        CommitmentKey(const size_t num_points, std::shared_ptr<barretenberg::srs::factories::CrsFactory> crs_factory);
+
+        Commitment commit(std::span<const Fr> polynomial);
+    };
+
+    class VerificationKey {
+      public:
+        barretenberg::scalar_multiplication::pippenger_runtime_state<curve::BN254> pippenger_runtime_state;
+        std::shared_ptr<barretenberg::srs::factories::ProverCrs<curve::BN254>> srs;
+
+        VerificationKey() = delete;
+
+        VerificationKey(size_t num_points, std::shared_ptr<barretenberg::srs::factories::CrsFactory> crs_factory);
+    };
+};
+
+} // namespace ipa
 
 namespace fake {
 
@@ -159,75 +169,5 @@ template <typename G> struct Params {
     };
 };
 } // namespace fake
-
-namespace ipa {
-
-struct Params {
-    using Fr = typename barretenberg::g1::Fr;
-    using Commitment = typename barretenberg::g1::affine_element;
-    using GroupElement = barretenberg::g1::element;
-
-    using Polynomial = barretenberg::Polynomial<Fr>;
-
-    class CommitmentKey;
-    class VerificationKey;
-
-    class CommitmentKey {
-
-      public:
-        CommitmentKey() = delete;
-
-        /**
-         * @brief Construct a new IPA Commitment Key object from existing SRS..
-         *
-         * @param num_points
-         * @param path
-         *
-         */
-        CommitmentKey(const size_t num_points, std::shared_ptr<barretenberg::srs::factories::CrsFactory> crs_factory)
-            : pippenger_runtime_state(num_points)
-            , srs(crs_factory->get_prover_crs(num_points))
-        {}
-
-        /**
-         * @brief Uses the ProverSRS to create an unblinded commitment to p(X)
-         *
-         * @param polynomial a univariate polynomial p(X) = ∑ᵢ aᵢ⋅Xⁱ ()
-         * @return Commitment computed as C = [p(x)] = ∑ᵢ aᵢ⋅Gᵢ where Gᵢ is the i-th element of the SRS
-         */
-        Commitment commit(std::span<const Fr> polynomial)
-        {
-            const size_t degree = polynomial.size();
-            ASSERT(degree <= srs->get_monomial_size());
-            return barretenberg::scalar_multiplication::pippenger_unsafe<curve::BN254>(
-                const_cast<Fr*>(polynomial.data()), srs->get_monomial_points(), degree, pippenger_runtime_state);
-        };
-
-        barretenberg::scalar_multiplication::pippenger_runtime_state<curve::BN254> pippenger_runtime_state;
-        std::shared_ptr<barretenberg::srs::factories::ProverCrs<curve::BN254>> srs;
-    };
-
-    class VerificationKey {
-      public:
-        VerificationKey() = delete;
-
-        /**
-         * @brief Construct a new IPA Verification Key object from existing SRS
-         *
-         *
-         * @param num_points specifies the length of the SRS
-         * @param path is the location to the SRS file
-         */
-        VerificationKey(size_t num_points, std::shared_ptr<barretenberg::srs::factories::CrsFactory> crs_factory)
-            : pippenger_runtime_state(num_points)
-            , srs(crs_factory->get_prover_crs(num_points))
-        {}
-
-        barretenberg::scalar_multiplication::pippenger_runtime_state<curve::BN254> pippenger_runtime_state;
-        std::shared_ptr<barretenberg::srs::factories::ProverCrs<curve::BN254>> srs;
-    };
-};
-
-} // namespace ipa
 
 } // namespace proof_system::pcs
